Extract RepositoryCallback from SessionRepository
Extracting the private class allows it to be reused in other *Repository classes.
